@@ -6,33 +6,35 @@
     "homepage": "https://github.com/joomla-framework/http",
     "license": "GPL-2.0-or-later",
     "require": {
-        "php": "^5.3.10|~7.0",
-        "joomla/uri": "~1.0|~2.0",
-        "composer/ca-bundle": "~1.0"
+        "php": "^7.2.5|^8.0",
+        "joomla/uri": "^1.0|^2.0",
+        "composer/ca-bundle": "^1.2.11",
+        "laminas/laminas-diactoros": "^2.2.2",
+        "psr/http-client": "^1.0",
+        "psr/http-message": "^1.0"
     },
     "require-dev": {
-        "joomla/test": "~1.0",
-        "joomla/coding-standards": "~2.0@alpha",
-        "phpunit/phpunit": "^4.8.35|^5.4.3|~6.0"
+        "joomla/coding-standards": "^2.0@alpha",
+        "joomla/test": "^2.0",
+        "phpunit/phpunit": "^8.5|^9.0"
     },
     "suggest": {
-        "joomla/registry": "Registry can be used as an alternative to using an array for the package options.",
         "ext-curl": "To use cURL for HTTP connections"
     },
     "autoload": {
-    	"psr-4": {
+        "psr-4": {
             "Joomla\\Http\\": "src/"
-    	}
+        }
     },
     "autoload-dev": {
-    	"psr-4": {
+        "psr-4": {
             "Joomla\\Http\\Tests\\": "Tests/"
-    	}
+        }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "extra": {
         "branch-alias": {
-            "dev-master": "1.x-dev"
+            "dev-2.0-dev": "2.0-dev"
         }
     }
 }
