@@ -1,6 +1,6 @@
 <?php
 /**
- * @copyright  Copyright (C) 2005 - 2019 Open Source Matters, Inc. All rights reserved.
+ * @copyright  Copyright (C) 2005 - 2021 Open Source Matters, Inc. All rights reserved.
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
@@ -15,12 +15,12 @@
 
 $response = new stdClass;
 
-$response->method = getVar($_SERVER, 'REQUEST_METHOD');
-$response->http_user_agent = getVar($_SERVER, 'HTTP_USER_AGENT');
-$response->request_uri = getVar($_SERVER, 'REQUEST_URI');
-$response->query_string = getVar($_SERVER, 'QUERY_STRING');
-$response->http_accept = getVar($_SERVER, 'HTTP_ACCEPT');
-$response->http_accept_charset = getVar($_SERVER, 'HTTP_ACCEPT_CHARSET');
+$response->method               = getVar($_SERVER, 'REQUEST_METHOD');
+$response->http_user_agent      = getVar($_SERVER, 'HTTP_USER_AGENT');
+$response->request_uri          = getVar($_SERVER, 'REQUEST_URI');
+$response->query_string         = getVar($_SERVER, 'QUERY_STRING');
+$response->http_accept          = getVar($_SERVER, 'HTTP_ACCEPT');
+$response->http_accept_charset  = getVar($_SERVER, 'HTTP_ACCEPT_CHARSET');
 $response->http_accept_encoding = getVar($_SERVER, 'HTTP_ACCEPT_ENCODING');
 
 $response->username = getVar($_SERVER, 'PHP_AUTH_USER');
@@ -28,13 +28,12 @@ $response->password = getVar($_SERVER, 'PHP_AUTH_PW');
 
 $response->http_referer = getVar($_SERVER, 'HTTP_REFERER');
 
-$response->get = $_GET;
-$response->post = $_POST;
-$response->files = $_FILES;
+$response->get     = $_GET;
+$response->post    = $_POST;
+$response->files   = $_FILES;
 $response->cookies = $_COOKIE;
 
 echo json_encode($response);
-
 
 /**
  * Retrieves a value from an array, returning a default value if not present
@@ -43,11 +42,9 @@ echo json_encode($response);
  * @param   string  $key      The value to retrieve.
  * @param   mixed   $default  The value to return if the key isn't present.
  *
- * @return  Mixed
- *
- * @since   11.4
+ * @return  mixed
  */
 function getVar($array, $key, $default = '')
 {
-	return isset($array[$key]) ? $array[$key] : $default;
+	return $array[$key] ?? $default;
 }
