@@ -2,24 +2,25 @@
 /**
  * Part of the Joomla Framework Http Package
  *
- * @copyright  Copyright (C) 2005 - 2019 Open Source Matters, Inc. All rights reserved.
+ * @copyright  Copyright (C) 2005 - 2021 Open Source Matters, Inc. All rights reserved.
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
 namespace Joomla\Http\Transport;
 
+use Joomla\Http\AbstractTransport;
 use Joomla\Http\Exception\InvalidResponseCodeException;
 use Joomla\Http\Response;
-use Joomla\Http\TransportInterface;
 use Joomla\Uri\Uri;
 use Joomla\Uri\UriInterface;
+use Laminas\Diactoros\Stream as StreamResponse;
 
 /**
  * HTTP transport class for using sockets directly.
  *
  * @since  1.0
  */
-class Socket implements TransportInterface
+class Socket extends AbstractTransport
 {
 	/**
 	 * Reusable socket connections.
@@ -28,39 +29,6 @@ class Socket implements TransportInterface
 	 * @since  1.0
 	 */
 	protected $connections;
-
-	/**
-	 * The client options.
-	 *
-	 * @var    array|\ArrayAccess
-	 * @since  1.0
-	 */
-	protected $options;
-
-	/**
-	 * Constructor.
-	 *
-	 * @param   array|\ArrayAccess  $options  Client options array.
-	 *
-	 * @since   1.0
-	 * @throws  \RuntimeException
-	 */
-	public function __construct($options = array())
-	{
-		if (!self::isSupported())
-		{
-			throw new \RuntimeException('Cannot use a socket transport when fsockopen() is not available.');
-		}
-
-		if (!\is_array($options) && !($options instanceof \ArrayAccess))
-		{
-			throw new \InvalidArgumentException(
-				'The options param must be an array or implement the ArrayAccess interface.'
-			);
-		}
-
-		$this->options = $options;
-	}
 
 	/**
 	 * Send a request to the server and return a Response object with the response.
@@ -77,28 +45,26 @@ class Socket implements TransportInterface
 	 * @since   1.0
 	 * @throws  \RuntimeException
 	 */
-	public function request($method, UriInterface $uri, $data = null, array $headers = null, $timeout = null, $userAgent = null)
+	public function request($method, UriInterface $uri, $data = null, array $headers = [], $timeout = null, $userAgent = null)
 	{
 		$connection = $this->connect($uri, $timeout);
 
 		// Make sure the connection is alive and valid.
-		if (\is_resource($connection))
-		{
-			// Make sure the connection has not timed out.
-			$meta = stream_get_meta_data($connection);
-
-			if ($meta['timed_out'])
-			{
-				throw new \RuntimeException('Server connection timed out.');
-			}
-		}
-		else
+		if (!\is_resource($connection))
 		{
 			throw new \RuntimeException('Not connected to server.');
 		}
 
+		// Make sure the connection has not timed out.
+		$meta = stream_get_meta_data($connection);
+
+		if ($meta['timed_out'])
+		{
+			throw new \RuntimeException('Server connection timed out.');
+		}
+
 		// Get the request path from the URI object.
-		$path = $uri->toString(array('path', 'query'));
+		$path = $uri->toString(['path', 'query']);
 
 		// If we have data to send make sure our request is setup for it.
 		if (!empty($data))
@@ -119,10 +85,10 @@ class Socket implements TransportInterface
 		}
 
 		// Configure protocol version, use transport's default if not set otherwise.
-		$protocolVersion = isset($this->options['protocolVersion']) ? $this->options['protocolVersion'] : '1.1';
+		$protocolVersion = $this->getOption('protocolVersion', '1.1');
 
 		// Build the request payload.
-		$request   = array();
+		$request   = [];
 		$request[] = strtoupper($method) . ' ' . ((empty($path)) ? '/' : $path) . ' HTTP/' . $protocolVersion;
 
 		if (!isset($headers['Host']))
@@ -144,27 +110,34 @@ class Socket implements TransportInterface
 		}
 
 		// If there are custom headers to send add them to the request payload.
-		if (\is_array($headers))
+		if (!empty($headers))
 		{
-			foreach ($headers as $k => $v)
+			foreach ($headers as $key => $value)
 			{
-				$request[] = $k . ': ' . $v;
+				if (\is_array($value))
+				{
+					foreach ($value as $header)
+					{
+						$request[] = "$key: $header";
+					}
+				}
+				else
+				{
+					$request[] = "$key: $value";
+				}
 			}
 		}
 
 		// Authentication, if needed
-		if (isset($this->options['userauth'], $this->options['passwordauth']))
+		if ($this->getOption('userauth') && $this->getOption('passwordauth'))
 		{
-			$request[] = 'Authorization: Basic ' . base64_encode($this->options['userauth'] . ':' . $this->options['passwordauth']);
+			$request[] = 'Authorization: Basic ' . base64_encode($this->getOption('userauth') . ':' . $this->getOption('passwordauth'));
 		}
 
 		// Set any custom transport options
-		if (isset($this->options['transport.socket']))
+		foreach ($this->getOption('transport.socket', []) as $value)
 		{
-			foreach ($this->options['transport.socket'] as $value)
-			{
-				$request[] = $value;
-			}
+			$request[] = $value;
 		}
 
 		// If we have data to send add it to the request payload.
@@ -188,9 +161,9 @@ class Socket implements TransportInterface
 		$content = $this->getResponse($content);
 
 		// Follow Http redirects
-		if ($content->code >= 301 && $content->code < 400 && isset($content->headers['Location']))
+		if ($content->getStatusCode() >= 301 && $content->getStatusCode() < 400 && $content->hasHeader('Location'))
 		{
-			return $this->request($method, new Uri($content->headers['Location']), $data, $headers, $timeout, $userAgent);
+			return $this->request($method, new Uri($content->getHeaderLine('Location')), $data, $headers, $timeout, $userAgent);
 		}
 
 		return $content;
@@ -209,9 +182,6 @@ class Socket implements TransportInterface
 	 */
 	protected function getResponse($content)
 	{
-		// Create the response object.
-		$return = new Response;
-
 		if (empty($content))
 		{
 			throw new \UnexpectedValueException('No content in response.');
@@ -224,30 +194,25 @@ class Socket implements TransportInterface
 		$headers = explode("\r\n", $response[0]);
 
 		// Set the body for the response.
-		$return->body = empty($response[1]) ? '' : $response[1];
+		$body = empty($response[1]) ? '' : $response[1];
 
 		// Get the response code from the first offset of the response headers.
 		preg_match('/[0-9]{3}/', array_shift($headers), $matches);
 		$code = $matches[0];
 
-		if (is_numeric($code))
-		{
-			$return->code = (int) $code;
-		}
-		else
+		if (!is_numeric($code))
 		{
 			// No valid response code was detected.
 			throw new InvalidResponseCodeException('No HTTP response code found.');
 		}
 
-		// Add the response headers to the response object.
-		foreach ($headers as $header)
-		{
-			$pos                                             = strpos($header, ':');
-			$return->headers[trim(substr($header, 0, $pos))] = trim(substr($header, ($pos + 1)));
-		}
+		$statusCode      = (int) $code;
+		$verifiedHeaders = $this->processHeaders($headers);
 
-		return $return;
+		$streamInterface = new StreamResponse('php://memory', 'rw');
+		$streamInterface->write($body);
+
+		return new Response($streamInterface, $statusCode, $verifiedHeaders);
 	}
 
 	/**
@@ -311,22 +276,7 @@ class Socket implements TransportInterface
 		}
 
 		// Capture PHP errors
-		if (PHP_VERSION_ID < 70000)
-		{
-			// @Todo Remove this path, when PHP5 support is dropped.
-			set_error_handler(
-				function () {
-					return false;
-				}
-			);
-			@trigger_error('');
-			restore_error_handler();
-		}
-		else
-		{
-			/** @noinspection PhpElementIsNotAvailableInCurrentPhpVersionInspection */
-			error_clear_last();
-		}
+		error_clear_last();
 
 		// PHP sends a warning if the uri does not exists; we silence it and throw an exception instead.
 		// Attempt to connect to the server
@@ -350,11 +300,7 @@ class Socket implements TransportInterface
 		// Since the connection was successful let's store it in case we need to use it later.
 		$this->connections[$key] = $connection;
 
-		// If an explicit timeout is set, set it.
-		if (isset($timeout))
-		{
-			stream_set_timeout($this->connections[$key], (int) $timeout);
-		}
+		stream_set_timeout($this->connections[$key], (int) $timeout);
 
 		return $this->connections[$key];
 	}
