@@ -2,20 +2,21 @@
 /**
  * Part of the Joomla Framework Http Package
  *
- * @copyright  Copyright (C) 2005 - 2019 Open Source Matters, Inc. All rights reserved.
+ * @copyright  Copyright (C) 2005 - 2021 Open Source Matters, Inc. All rights reserved.
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
 namespace Joomla\Http\Exception;
 
 use Joomla\Http\Response;
+use Psr\Http\Client\ClientExceptionInterface;
 
 /**
  * Exception representing an unexpected response
  *
  * @since  1.2.0
  */
-class UnexpectedResponseException extends \DomainException
+class UnexpectedResponseException extends \DomainException implements ClientExceptionInterface
 {
 	/**
 	 * The Response object.
