@@ -2,7 +2,7 @@
 /**
  * Part of the Joomla Framework Http Package
  *
- * @copyright  Copyright (C) 2005 - 2019 Open Source Matters, Inc. All rights reserved.
+ * @copyright  Copyright (C) 2005 - 2021 Open Source Matters, Inc. All rights reserved.
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
@@ -18,15 +18,6 @@ use Joomla\Uri\UriInterface;
 interface TransportInterface
 {
 	/**
-	 * Constructor.
-	 *
-	 * @param   array|\ArrayAccess  $options  Client options object.
-	 *
-	 * @since   1.0
-	 */
-	public function __construct($options = array());
-
-	/**
 	 * Send a request to the server and return a Response object with the response.
 	 *
 	 * @param   string        $method     The HTTP method for sending the request.
@@ -40,7 +31,7 @@ interface TransportInterface
 	 *
 	 * @since   1.0
 	 */
-	public function request($method, UriInterface $uri, $data = null, array $headers = null, $timeout = null, $userAgent = null);
+	public function request($method, UriInterface $uri, $data = null, array $headers = [], $timeout = null, $userAgent = null);
 
 	/**
 	 * Method to check if HTTP transport layer available for using
